requirements.txt
asgiref==3.3.4
attrs==20.3.0
autopep8==1.5.6
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
djangorestframework==3.12.4
drf-extensions==0.7.0
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==1.1.1
packaging==20.9
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
swagger-spec-validator==2.7.3
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
uWSGI==2.0.19.1
